+++ react/country/AUT.ts
@@ -98,6 +98,7 @@ const rules: PostalCodeRules = {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: false,
     },
 
     number: {