+++ react/country/POL.ts
@@ -105,6 +105,7 @@ const rules: PostalCodeRules = {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: false,
     },
 
     number: {