+++ react/country/ROU.ts
@@ -14025,6 +14025,7 @@ const rules: PostalCodeRules = {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: false,
     },
     number: {
       valueIn: 'long_name',