+++ react/country/VEN.ts
@@ -291,6 +291,7 @@ const rules: PostalCodeRules = {
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],
+      required: false,
     },
 
     complement: {